updating validator.py file location
