+++ definition.py
@@ -1,4 +1,4 @@
-import os 
+import os
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-CONFIG_PATH = os.path.join(ROOT_DIR,'config/logging.conf')
+CONFIG_PATH = os.path.join(ROOT_DIR, 'config/logging.conf')
