+++ src/com/jalasoft/search_files/utils/logging.py
@@ -8,3 +8,4 @@ logger = logging.getLogger('SearchFiles')
 
 # Set default logger's level
 logger.setLevel(logging.INFO)
+logger.setLevel(logging.ERROR)
